Summary of Changes
  Files with no changes needed (already excellent)

  - ParallelMode.java, PrimitiveTuple.java, ImmutableIntArray.java,
Points.java,
   Matrices.java — already had thorough, JDK-compliant Javadocs

  Consistency fixes across all 9 matrix classes + Matrix.java

  ┌────────────────────────────────────┬────────────────────────────────────┐
  │                Fix                 │           Files affected
│
  ├────────────────────────────────────┼────────────────────────────────────┤
  │                                    │ BooleanMatrix, ByteMatrix,
│
  │ 1xsize → {@code 1 x size} in       │ CharMatrix, DoubleMatrix,
│
  │ random(int)                        │ FloatMatrix, IntMatrix,
│
  │                                    │ LongMatrix, ShortMatrix
│
  ├────────────────────────────────────┼────────────────────────────────────┤
  │ "row/column number" → "row/column  │ All 9 primitive matrix classes
│
  │ count" in resize()                 │
│
  ├────────────────────────────────────┼────────────────────────────────────┤
  │ println() summary standardized to  │
│
  │ "Prints this matrix to standard    │ All 10 matrix classes
│
  │ output and returns the formatted   │
│
  │ string."                           │
│
  ├────────────────────────────────────┼────────────────────────────────────┤
  │ "maybe" → "may be" typo in         │ CharMatrix, DoubleMatrix,
│
  │ copyFrom()                         │ FloatMatrix
│
  └────────────────────────────────────┴────────────────────────────────────┘

  Arrays.java

  - Fixed ~25+ bare null references → {@code null} in @return and @param
tags

  Matrix.java

  - Added {@link ...} to ~20 @return tags referencing types (Nullable,
  BooleanMatrix, ByteMatrix, CharMatrix, ShortMatrix, IntMatrix,
LongMatrix,
  FloatMatrix, DoubleMatrix, Stream, DataSet)

  FloatMatrix.java

  - Added 2 missing @throws IndexOutOfBoundsException on
streamHorizontal(int)
  and streamVertical(int)

  IntMatrix.java

  - Added 2 missing @throws IllegalArgumentException on mainDiagonal()
and
  antiDiagonal()

  ShortMatrix.java & LongMatrix.java

  - Added usage examples and @see tags on rowCopy() methods

  LongMatrix.java

  - Added {@code null} note to length() @param tag

  5 Tuple files (DoubleTuple, FloatTuple, IntTuple, LongTuple,
ShortTuple)

  - Added 50 missing Javadoc comments on protected elements() overrides
(10
  inner classes × 5 files)

+++ src/test/java/com/landawn/abacus/util/ByteMatrixTest.java
@@ -9,18 +9,21 @@ import static org.junit.jupiter.api.Assertions.assertNotSame;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import com.landawn.abacus.TestBase;
-import com.landawn.abacus.util.Sheet.Point;
-import com.landawn.abacus.util.stream.ByteStream;
-import com.landawn.abacus.util.stream.Stream;
-import com.landawn.abacus.util.u.OptionalByte;
+
 import java.util.ArrayList;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
+
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+
+import com.landawn.abacus.TestBase;
+import com.landawn.abacus.util.Sheet.Point;
+import com.landawn.abacus.util.u.OptionalByte;
+import com.landawn.abacus.util.stream.ByteStream;
+import com.landawn.abacus.util.stream.Stream;
 
 class ByteMatrixTest extends TestBase {
 
@@ -4168,7 +4171,7 @@ class ByteMatrixTest extends TestBase {
         @Test
         public void testMapToObj_withComplexType() {
             ByteMatrix m = ByteMatrix.of(new byte[][] { { 1, 2 }, { 3, 4 } });
-            Matrix<Integer> mapped = m.mapToObj(val -> (int) val * 10, Integer.class);
+            Matrix<Integer> mapped = m.mapToObj(val -> val * 10, Integer.class);
             assertEquals(10, mapped.get(0, 0));
             assertEquals(20, mapped.get(0, 1));
             assertEquals(30, mapped.get(1, 0));
@@ -4605,7 +4608,7 @@ class ByteMatrixTest extends TestBase {
         public void testForEach_withConsumer() {
             ByteMatrix m = ByteMatrix.of(new byte[][] { { 1, 2 }, { 3, 4 } });
             AtomicInteger sum = new AtomicInteger(0);
-            m.forEach(val -> sum.addAndGet((int) val));
+            m.forEach(val -> sum.addAndGet(val));
             assertEquals(10, sum.get()); // 1+2+3+4 = 10
         }
